+++ M2/Macaulay2/e/aring-gf.hpp
@@ -179,7 +179,7 @@ class ARingGF : public RingInterface
         @{ */
     void to_ring_elem(ring_elem &result, const ElementType &a) const
     {
-      result.int_val = a;
+      result.int_val = static_cast<int>(a);
     }
  
     void from_ring_elem(ElementType &result, const ring_elem &a) const
